+++ tests/test_data/shell_scripts/job.sh
@@ -1,0 +1,5 @@
+
+printf "hello, this is from shell job scripts \n"
+
+printf "passing arguments is "
+printf "$@"